+++ packages/storybook-angular/src/lib/start-storybook/start-storybook.ts
@@ -1,0 +1,3 @@
+import StartStorybookBuilder from '@storybook/angular/dist/builders/start-storybook/index.js';
+
+export default StartStorybookBuilder;