+++ tools/publish.sh
@@ -5,6 +5,7 @@ npm publish node_modules/@analogjs/astro-angular --access public --tag $RELEASE_
 npm publish node_modules/@analogjs/content --access public --tag $RELEASE_TAG
 npm publish node_modules/@analogjs/platform --access public --tag $RELEASE_TAG
 npm publish node_modules/@analogjs/router --access public --tag $RELEASE_TAG
+npm publish node_modules/@analogjs/storybook-angular --access public --tag $RELEASE_TAG
 npm publish node_modules/@analogjs/vite-plugin-angular --access public --tag $RELEASE_TAG
 npm publish node_modules/@analogjs/vite-plugin-nitro --access public --tag $RELEASE_TAG
 npm publish node_modules/@analogjs/vitest-angular --access public --tag $RELEASE_TAG